move coll residuals getter function, simplify ocp code
