+++ agimus_controller/agimus_controller/ocp/ocp_traj_tracking_collision_avoidance.py
@@ -2,6 +2,8 @@ import crocoddyl
 import numpy as np
 import pinocchio as pin
 from typing import Tuple
+from typing import Union
+import numpy.typing as npt
 
 from agimus_controller.ocp_base_croco import OCPBaseCroco
 from agimus_controller.trajectory import WeightedTrajectoryPoint
@@ -38,60 +40,56 @@ class OCPCrocoTrajTrackCollAvoidance(OCPBaseCroco):
 
     def create_running_model_list(self) -> list[crocoddyl.ActionModelAbstract]:
         running_model_list = []
-        for factor_idx, dts in enumerate(self._params.dt_factor_n_seq.dts):
-            factor = self._params.dt_factor_n_seq.factors[factor_idx]
-            dt = factor * self._params.dt
-            for _ in range(dts):
-                running_cost_model = crocoddyl.CostModelSum(self._state)
-                x_residual, u_residual, frame_placement_residual = (
-                    self.create_residuals()
+        running_cost_model = crocoddyl.CostModelSum(self._state)
+        x_residual, u_residual, frame_placement_residual = self.create_residuals()
+        x_reg_cost = crocoddyl.CostModelResidual(
+            self._state,
+            crocoddyl.ActivationModelWeightedQuad(
+                np.ones(
+                    self._robot_models.robot_model.nq
+                    + self._robot_models.robot_model.nv
                 )
-                x_reg_cost = crocoddyl.CostModelResidual(
-                    self._state,
-                    crocoddyl.ActivationModelWeightedQuad(
-                        np.ones(
-                            self._robot_models.robot_model.nq
-                            + self._robot_models.robot_model.nv
-                        )
-                    ),
-                    x_residual,
-                )
-                u_reg_cost = crocoddyl.CostModelResidual(
-                    self._state,
-                    crocoddyl.ActivationModelWeightedQuad(
-                        np.ones(self._robot_models.robot_model.nv)
-                    ),
-                    u_residual,
-                )
-                ee_pose_cost = crocoddyl.CostModelResidual(
-                    self._state,
-                    crocoddyl.ActivationModelWeightedQuad(np.ones(6)),
-                    frame_placement_residual,
-                )
-                for col_idx in range(len(self._collision_model.collisionPairs)):
-                    activation = ActivationModelQuadExp(
-                        1, self._params.activation_distance_threshold**2
-                    )
-                    distance_coll_residual = ResidualDistanceCollision(
-                        self._state, 7, self._collision_model, col_idx
-                    )
-                    distance_coll_cost = crocoddyl.CostModelResidual(
-                        self._state, activation, distance_coll_residual
-                    )
-                    running_cost_model.addCost(
-                        f"distColl{col_idx}", distance_coll_cost, 1.0
-                    )
-                running_cost_model.addCost("stateReg", x_reg_cost, 1.0)
-                running_cost_model.addCost("ctrlReg", u_reg_cost, 1.0)
-                running_cost_model.addCost("goalTracking", ee_pose_cost, 1.0)
+            ),
+            x_residual,
+        )
+        u_reg_cost = crocoddyl.CostModelResidual(
+            self._state,
+            crocoddyl.ActivationModelWeightedQuad(
+                np.ones(self._robot_models.robot_model.nv)
+            ),
+            u_residual,
+        )
+        ee_pose_cost = crocoddyl.CostModelResidual(
+            self._state,
+            crocoddyl.ActivationModelWeightedQuad(np.ones(6)),
+            frame_placement_residual,
+        )
+        for col_idx in range(len(self._collision_model.collisionPairs)):
+            activation = ActivationModelQuadExp(
+                1, self._params.activation_distance_threshold**2
+            )
+            distance_coll_residual = ResidualDistanceCollision(
+                self._state, 7, self._collision_model, col_idx
+            )
+            distance_coll_cost = crocoddyl.CostModelResidual(
+                self._state, activation, distance_coll_residual
+            )
+            running_cost_model.addCost(f"distColl{col_idx}", distance_coll_cost, 1.0)
+        running_cost_model.addCost("stateReg", x_reg_cost, 1.0)
+        running_cost_model.addCost("ctrlReg", u_reg_cost, 1.0)
+        running_cost_model.addCost("goalTracking", ee_pose_cost, 1.0)
 
-                constraints = self.get_collision_constraints(active=True)
-                running_DAM = crocoddyl.DifferentialActionModelFreeFwdDynamics(
-                    self._state, self._actuation, running_cost_model, constraints
-                )
-                running_model = crocoddyl.IntegratedActionModelEuler(running_DAM, dt)
-                running_model.differential.armature = self._robot_models.armature
-                running_model_list.append(running_model)
+        constraints = self._get_collision_constraints(active=True)
+        running_DAM = crocoddyl.DifferentialActionModelFreeFwdDynamics(
+            self._state, self._actuation, running_cost_model, constraints
+        )
+        for factor, dts in zip(
+            self._params.dt_factor_n_seq.factors, self._params.dt_factor_n_seq.dts
+        ):
+            dt = factor * self._params.dt
+            running_model = crocoddyl.IntegratedActionModelEuler(running_DAM, dt)
+            running_model.differential.armature = self._robot_models.armature
+            running_model_list.extend([running_model] * dts)
         return running_model_list
 
     def create_terminal_model(self) -> crocoddyl.ActionModelAbstract:
@@ -124,42 +122,65 @@ class OCPCrocoTrajTrackCollAvoidance(OCPBaseCroco):
             terminal_cost_model.addCost(f"distColl{col_idx}", distance_coll_cost, 1.0)
         terminal_cost_model.addCost("stateReg", x_reg_cost, 1.0)
         terminal_cost_model.addCost("goalTracking", ee_pose_cost, 1.0)
-        constraints = self.get_collision_constraints(active=True)
+        constraints = self._get_collision_constraints(active=True)
         terminal_DAM = crocoddyl.DifferentialActionModelFreeFwdDynamics(
             self._state, self._actuation, terminal_cost_model, constraints
         )
 
-        terminal_model = crocoddyl.IntegratedActionModelEuler(terminal_DAM, 0.0)
+        terminal_model = crocoddyl.IntegratedActionModelEuler(terminal_DAM, 1.0)
         terminal_model.differential.armature = self._robot_models.armature
         return terminal_model
 
-    def get_collision_constraints(self, active=True):
+    def _get_collision_constraints(self, active=True):
         """Returns the collision constraints model manager."""
         constraint_model_manager = crocoddyl.ConstraintModelManager(
             self._state, self.nq
         )
-        if len(self._collision_model.collisionPairs) != 0:
-            for col_idx in range(len(self._collision_model.collisionPairs)):
-                collision_constraint = self._get_collision_constraint_residual(col_idx)
-                # Adding the constraint to the constraint manager
-                constraint_model_manager.addConstraint(
-                    "col_term_" + str(col_idx), collision_constraint, active=active
-                )
+        for col_idx in range(len(self._collision_model.collisionPairs)):
+            distance_coll_residual = ResidualDistanceCollision(
+                self._state, 7, self._collision_model, col_idx
+            )
+            collision_constraint = crocoddyl.ConstraintModelResidual(
+                self._state,
+                distance_coll_residual,
+                np.array([self._params.collision_safety_margin]),
+                np.array([np.inf]),
+            )
+            # Adding the constraint to the constraint manager
+            constraint_model_manager.addConstraint(
+                "col_term_" + str(col_idx), collision_constraint, active=active
+            )
         return constraint_model_manager
 
-    def _get_collision_constraint_residual(
-        self, col_idx: int
-    ) -> "crocoddyl.ConstraintModelResidual":
-        """Returns the collision constraint model residual."""
-        distance_coll_residual = ResidualDistanceCollision(
-            self._state, 7, self._collision_model, col_idx
-        )
-        return crocoddyl.ConstraintModelResidual(
-            self._state,
-            distance_coll_residual,
-            np.array([self._params.collision_safety_margin]),
-            np.array([np.inf]),
-        )
+    def get_distance_collision_residuals(self) -> Union[npt.NDArray[np.float64], None]:
+        """Return distance collision residual if Crocoddyl's problem use it."""
+        nb_collision_pairs = len(self._collision_model.collisionPairs)
+        if (
+            nb_collision_pairs != 0
+            and self._solver.problem.runningDatas[0].differential.constraints
+            is not None
+        ):
+            coll_residuals = np.zeros((self._params.horizon_size, nb_collision_pairs))
+            for node_idx in range(self._params.horizon_size - 1):
+                constraints_residual_dict = self._solver.problem.runningDatas[
+                    node_idx
+                ].differential.constraints.constraints.todict()
+                for coll_pair_idx, constraint_key in enumerate(
+                    constraints_residual_dict.keys()
+                ):
+                    coll_residuals[node_idx, coll_pair_idx] = constraints_residual_dict[
+                        constraint_key
+                    ].residual.r[0]
+            constraints_residual_dict = self._solver.problem.terminalData.differential.constraints.constraints.todict()
+            for coll_pair_idx, constraint_key in enumerate(
+                constraints_residual_dict.keys()
+            ):
+                coll_residuals[self._params.horizon_size - 1, coll_pair_idx] = (
+                    constraints_residual_dict[constraint_key].residual.r[0]
+                )
+            return coll_residuals
+        else:
+            return None
 
     def set_reference_weighted_trajectory(
         self, reference_weighted_trajectory: list[WeightedTrajectoryPoint]
@@ -191,7 +212,7 @@ class OCPCrocoTrajTrackCollAvoidance(OCPBaseCroco):
             self.modify_cost_reference_and_weights(model, "ctrlReg", u_ref, u_weights)
 
             ee_names = list(
-                iter(reference_weighted_trajectory[i].weights.w_end_effector_poses)
+                reference_weighted_trajectory[i].weights.w_end_effector_poses.keys()
             )
             if len(ee_names) > 1:
                 raise ValueError("Only one end-effector tracking reference is allowed.")
@@ -227,7 +248,7 @@ class OCPCrocoTrajTrackCollAvoidance(OCPBaseCroco):
         )
         # Modify end effector frame cost
         ee_names = list(
-            iter(reference_weighted_trajectory[-1].weights.w_end_effector_poses)
+            reference_weighted_trajectory[-1].weights.w_end_effector_poses.keys()
         )
         if len(ee_names) > 1:
             raise ValueError("Only one end-effector tracking reference is allowed.")
